Included socket path and permission variables in docs.
The socket variables were not present in the README docs, and a user would not know they are options unless they check or stumble upon defaults/main.yml. Granted that most people will be using Redis over TCP, but many also use Redis as a local cache, so these settings should be present.