Recreate /var/run/redis after reboot.
In Ubuntu/Debian `/var/run` is a tmpfs, so all of its content is deleted upon reboot. Common practice is for init scripts is to recreate the service's directory under `/var/run` upon starting the service. This commit adds this necessary feature to the Redis init script, by checking if `/var/run/redis` exists upon start, if it doesn't, then the directory is created, ownership is granted to the `redis_user`, and appropriate permissions are applied.