Redis server and sentinel can be installed on the same host. Fixed systemd unit files for Debian/Ubuntu (tested on Ubuntu 20.04).
For systemd based services the PIDfile directory will not be created anymore, it will be created automatically by systemd via the 'RuntimeDirectory' setting.
Added a couple of configuration settings to the systemd unit file which were copied from the unit file which is installed on Ubuntu via the official repository (which is currently at major version 5 of redis, maybe other settings might be better for version 6 - nevertheless the service seems to work).